==> 03 Scripts/1. Descriptive analysis.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

path = r'C:\Users\kenta\Google ドライブ\CareerFoundry Immersion\Instacart Basket Analysis'

path

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orderss.csv'), usecols = vars_list)

df.head(5)

# import order data
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_prods.head(20)

df_prods.tail(35)

df_prods.columns

df_prods.shape

df_prods.describe()
# What’s the maximum value of the “aisle_id” column? A 134,00

df_prods.dtypes
# What’s the data type of the “department_id” column? A Integer
