==> 03 Scripts/2. Data Wrangling.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

df_ords = pd.read_csv(r'C:\Users\kenta\Google ドライブ\CareerFoundry Immersion\Instacart Basket Analysis\02 Data\Original Data\orders.csv', index_col = False)

path = r'C:\Users\kenta\Google ドライブ\CareerFoundry Immersion\Instacart Basket Analysis'

# 2. Find another identifier variable in the df_ords dataframe that doesn’t need to be included in your analysis as a numeric variable and change it to a suitable format.
df_ords.drop(columns = ['eval_set'])

# 3. Look for a variable in your df_ords dataframe with an unintuitive name and change its name without overwriting the data frame.
df_ords_2 = df_ords.drop(columns = ['eval_set'])

# 4. Your client wants to know what the busiest hour is for placing orders. Find the frequency of the corresponding variable and share your findings.
df_ords_2['order_hour_of_day'].value_counts()

# 5. Determine the meaning behind a value of 4 in the "department_id" column within the df_prods dataframe using a data dictionary.
df_prods = pd.read_csv(r'C:\Users\kenta\Google ドライブ\CareerFoundry Immersion\Instacart Basket Analysis\02 Data\Original Data\products.csv', index_col = False)

df_dep = pd.read_csv(r'C:\Users\kenta\Google ドライブ\CareerFoundry Immersion\Instacart Basket Analysis\02 Data\Original Data\departments.csv', index_col = False)

df_dep.T
# A. 4 = produce

# 6. The sales team in your client’s organization wants to know more about breakfast item sales. Create a subset containing only the required information.
df_prods['department_id']==14

df_prods[df_prods['department_id']==14]

df_breakfast = df_prods[df_prods['department_id']==14]

# 7. They’d also like to see details about customers who might be throwing dinner parties. Your task is to find all observations from the entire dataframe that include items from the following departments: alcohol, deli, beverages, and meat/seafood. You’ll need to present this subset to your client.
df_prods.head(20)

df_prods[df_prods['department_id'].isin(['5', '7', '12', '20'])]

df_dinner_parties = df_prods[df_prods['department_id'].isin(['5', '7', '12', '20'])]

df_dinner_parties.describe()

# 8. It’s important that you keep track of total counts in your dataframes. How many rows does the last dataframe you created have?
# A 7650

# 9. Someone from the data engineers team in Instacart thinks they’ve spotted something strange about the customer with a "user_id" of “1.” Extract all the information you can about this user.
df_ords['user_id']==1

df_ords[df_ords['user_id']==1]

df_user1 = df_ords[df_ords['user_id']==1]

# 10. You also need to provide some details about this user’s behavior. What basic stats can you provide based on the information you have?
df_user1.describe()

# 12. Export your df_ords dataframe as “orders_wrangled.csv” in your “Prepared Data” folder.
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

df_dep.T

df_dep_t = df_dep.T

df_dep_t.reset_index()

new_header = df_dep_t.iloc[0]

df_dep_t_new = df_dep_t[1:]

df_dep_t_new.columns = new_header

df_dep_t_new

# 13, Export the df_dep_t_new dataframe as “departments_wrangled.csv” in your “Prepared Data” folder so that you have a “.csv” file of your departments data in the correct format.
df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.cs'))
