==> 03 Scripts/3. Cleaning data.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

df_prods = pd.read_csv(r'C:\Users\kenta\Google ドライブ\CareerFoundry Immersion\Instacart Basket Analysis\02 Data\Original Data\products.csv', index_col = False)

df_ords = pd.read_csv(r'C:\Users\kenta\Google ドライブ\CareerFoundry Immersion\Instacart Basket Analysis\02 Data\Original Data\orders.csv', index_col = False)

# 1.If you haven’t performed the consistency checks covered in this Exercise on your df_prods dataframe, do so now.
df_prods.describe()

2. Using your new knowledge about how to interpret the output of this function, share in a markdown cell whether anything about the data looks off or should be investigated further.

A. The Max cell in the Price column is extraordinary higher.

# 3. Check for mixed-type data in your df_ords dataframe. No mix type of data
for col in df_ords.columns.tolist():
    weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (df_ords[weird]) > 0:
        print (col)

# 5. Run a check for missing values in your df_ords dataframe.
df_ords.isnull().sum() 

df_ords.shape

pd.options.display.float_format = '{:.2f}'.format

df_ords.describe()

In a markdown cell, report your findings and propose an explanation for any missing values you find.

206209 data of days_since_prior_order is missing out of 3421083, which is the relatively high proportion of the column. Average days_since_prior_order is 11.11 days, the median is seven days, the max is 30, and the minimum is 0 day.

df_ords.head()

df_ords[df_ords['days_since_prior_order'].isnull()]

df_ords['order_number'].value_counts()

# 6. Address the missing values using an appropriate method.

In a markdown cell, explain why you used your method of choice.

I just left it as it is because all data that shows nan is their 1st order, which has no prior deal before the purchase.

# 7. Run a check for duplicate values in your df_ords data.
df_ords_dups = df_ords[df_ords.duplicated()]

df_ords_dups

No dupe

path = r'C:\Users\kenta\Google ドライブ\CareerFoundry Immersion\Instacart Basket Analysis'

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_cleansed.csv'))
df_prods.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_cleansed.csv'))
