==> 03 Scripts/4. Combining & Exporting Data 2.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

df_ords_prior = pd.read_csv(r'C:\Users\kenta\Instacart Basket Analysis\02 Data\Original Data\order_products__prior.csv', index_col = False)

df_ords = pd.read_csv(r'C:\Users\kenta\Instacart Basket Analysis\02 Data\Prepared Data\orders_cleansed.csv', index_col = False)

# If you haven’t merged your prepared Instacart orders data with the new orders_products_prior dataframe per the instructions in this Exercise, do so now.
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', how = 'inner', indicator = True) 

df_merged_large

df_merged_large['_merge'].value_counts()

# 2. Export the merged file in pickle format as “orders_products_combined.pkl”.
df_merged_large.to_pickle(r'C:\Users\kenta\Instacart Basket Analysis\02 Data\Prepared Data\orders_products_combined.pkl')

df_ords_prior.isnull().sum() 

df_ords_prior.duplicated().sum()
