==> 03 Scripts/4. Combining & Exporting Data.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

# 3. In a new notebook, import the orders_products_combined dataframe from the pickle file you just saved.
df_orders_products_combined = pd.read_pickle(r'C:\Users\kenta\Instacart Basket Analysis\02 Data\Prepared Data\orders_products_combined.pkl')

# 4. Check the shape of the imported dataframe (it should be the same as the one you exported—always check!).
df_orders_products_combined.head()

df_orders_products_combined.shape

df_product = pd.read_csv(r'C:\Users\kenta\Instacart Basket Analysis\02 Data\Prepared Data\df_prods_clean_no_dups.csv', index_col = False)

df_product.head()

df_product.shape

# remove unnecessary column
df_products = df_product.drop(columns = ['Unnamed: 0',])

df_products.head()

df_products.shape

df_orders_products_combined_1 = df_orders_products_combined.drop(columns = ['_merge','Unnamed: 0'])

# 5. Determine a suitable way to combine the orders_products_combined dataframe with your products data set. Make sure you’re using your wrangled, cleaned, and deduped products data set stored in your “Prepared Data” folder from the previous Exercise’s task.
orders_products_merged = df_orders_products_combined_1.merge(df_products, on = ['product_id'], how = 'left',indicator = True)

orders_products_merged.head()

orders_products_merged.shape

# 6. Confirm the results of the merge using the merge flag.
orders_products_merged['_merge'].value_counts()

orders_products_merged.to_pickle(r'C:\Users\kenta\Instacart Basket Analysis\02 Data\Prepared Data\orders_products_merged.pkl')
