==> 03 Scripts/7. Data Visualization.ipynb <==
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

customer = pd.read_csv(r'C:\Users\kenta\Instacart Basket Analysis\02 Data\Original Data\customers.csv',index_col = False)

# 4. Wrangle the data so that it follows consistent logic; for example, rename columns with illogical names and drop columns that don’t add anything to your analysis.
# rename the n_dependent to Number_of_dependents and fam_status to family_status
customer.rename(columns = {'n_dependants' : 'Number_of_dependents'}, inplace = True)

customer.rename(columns = {'fam_status' : 'family_status'}, inplace = True)

# 5. Complete the fundamental data quality and consistency checks you’ve learned throughout this Achievement; for example, check for and address missing values and duplicates, and convert any mixed-type data.
customer

pd.options.display.float_format = '{:.2f}'.format

customer.describe()

customer.shape

customer.info()

customer.isnull().sum() 

customer_nan = customer[customer['First Name'].isnull()== True]

# missing data
customer_nan

# Finding Duplicates
customer_dup = customer[customer.duplicated()]

customer_dup

# 6. Combine your customer data with the rest of your prepared Instacart data.
ords_prods_merge = pd.read_pickle(r'C:\Users\kenta\Instacart Basket Analysis\02 Data\Prepared Data\ords_prods_merge_after_grouping.pkl')

ords_prods_merge.head()

ords_prods_merge_1 = ords_prods_merge.drop(columns = ['_merge'])

ords_prods_merge_1.head()

customer_merged = ords_prods_merge_1.merge(customer, on = ['user_id'], how = 'left',indicator = True)

customer_merged.head()

customer_merged['_merge'].value_counts()

customer_merged.shape

customer_merged.to_pickle(r'C:\Users\kenta\Instacart Basket Analysis\02 Data\Prepared Data\customer_merged_with_grouped.pkl')
